Fix network charts in VM Groups Perf Analysis
Networks charts now utilize two different counters due to the difference
in the way Windows and Linux reports counters to VMConnections table. A
hidden parameter was added so network charts will be displayed if need
be. A disclaimer is shown when the user selects a network counter that
additional counters other than the one selected is being used.