Add correlation to user with long lookback of queries
Like in Anita, they used to show users that are running queries with long lookback. can maybe correlate between the id in commands and queries. 

Anita had the "Detailed View" Button.
It opens a grid (under the histogram) with columns for: User, application, Number of queries.
Once you clicked on a bar, the grid changed accordingly and showed the top users (by query count) who ran the queries for the selected lookback period (per day)