Fix aggregation for provisioned subscribers
Use maximum rather than average as the aggregation for provisioned
subscribers - bringing in line with the other metrics (connected nodeBs,
registered/connected subscribers). This wasn't possible previously due
to a metrics bug, but this is now fixed. This avoids an issue where
provisioned subscribers can show as a decimal value if there are
multiple values being aggregated over.