New Email tab in ACS Insights workbook
Incorporating the new Email tab to the main ACS Insights workbook. The elements are taken from the Email Workbook template.

For tighter integration with the rest of the Insights tab, the date range and time granularity parameters were removed from the Email group and queries were changed to use the global parameters.