Update Overview.workbook update datetime column (#2662)
We have an issue in our Azure Communications Email Service Insights Workbook where our per-instance table shows an aggregated date (bin()). The issue is in the query used to generate this chart. The Date Sent column is using the bin() function to round to the nearest time bucket. This instance level chart should use actual timestamps