Default AllocationId
If AllocationId not passed as URL parameter, then choose the AllocationId with the latest assignment event. This allows the bot to direct users to the latest analysis results for a feature flag, without being specific about the experiment (i.e., AllocationId).

This does not impact how users find analysis results from within the workbook. Clicking on an experiment in the Experiment Status tab will always set the AllocationId parameter value.